pyproject.toml
[project]
name = "django_academy"
version = "0.0.1"
dependencies = [
    "asgiref==3.6.0",
    "Django==3.2.16",
    "django-environ",
    "pytz==2022.7.1",
    "sqlparse==0.4.3",
    "django-debug-toolbar==3.8.1",
	"django-cleanup==7.0.0",
	"sorl-thumbnail==12.9.0",
	"django_compressor==4.3.1",
	"django-libsass==0.9",
	"pillow==9.4.0",
	"django-quill-editor==0.1.40"
]

[tool.setuptools]
packages = ["about", "catalog", "homepage", "myserver", "core"]

[project.optional-dependencies]
    dev = ["flake8==6.0.0", "black==23.1.0", "pep8-naming", 
	   "flake8-import-order","flake8-quotes", "Flake8-pyproject", "djlint==1.19.16"]
    test = ["pytest==7.2.1", "pytest-django==4.5.2"]

[tool.black]
line-length = 79
target-version = ["py310"]

[tool.pytest.ini_options]
minversion = "6.0"
DJANGO_SETTINGS_MODULE = "myserver.settings"

